+++ src/test/converter2/issues/gh2031.ts
@@ -1,0 +1,10 @@
+export class MyClass {
+    /**
+     * {@link anotherMethod}
+     *
+     * @param aParam {@link anotherMethod}
+     */
+    aMethod(aParam: number) {}
+
+    anotherMethod() {}
+}