+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Bug Fixes
 
 -   Readme files within monorepos now have `@link` tags resolved, #2029.
+-   Correctly resolve unqualified links to class members within parameters, #2031.
 
 ## v0.23.10 (2022-07-31)